Breaking up TicTacToe class
