+++ test_tictactoe.py
@@ -13,25 +13,6 @@ class TestDummy(unittest.TestCase):
 
 
 class TestTicTacToe(unittest.TestCase):
-    # Testing welcome message and game instructions
-    def test_welcome_message(self):
-        test_game = TicTacToe()
-        # Tests display of welcome message
-        captured_output = StringIO()
-        sys.stdout = captured_output
-        test_game.welcome_message()
-        output = captured_output.getvalue().strip()
-        self.assertEqual(output, "Welcome to Tic Tac Toe")
-
-    def test_game_instructions(self):
-        test_game = TicTacToe()
-        # Tests Display of Game Instructions
-        captured_output = StringIO()
-        sys.stdout = captured_output
-        test_game.game_instructions()
-        validate_instructions = "Each square on the board have a value from 1-9. Select which square you would like to play by inputting the correct value when promoted."
-        output = captured_output.getvalue().strip()
-        self.assertEqual(output, validate_instructions)
 
     # Initialise and Assign Players
     def test_create_player(self):
